:construction: [Add] error class
