--- utils/utils.rb
@@ -1,9 +1,0 @@
-# frozon_string_literal: true
-
-# クラス化してインスタンス化をするのであれば、何かモノとみなすということ
-# だからstaticなmethodの集まりにする。
-module Utils
-  class Util
-  
-  end
-end
